Remove filter evaluation
