+++ vortex-layout/src/layouts/mod.rs
@@ -14,7 +14,6 @@ pub mod compact;
 pub mod compressed;
 pub mod dict;
 pub mod file_stats;
-pub mod filter;
 pub mod flat;
 mod partitioned;
 pub mod repartition;
