+++ vortex-scan/src/filter.rs
@@ -1,0 +1,110 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use bit_vec::BitVec;
+use itertools::Itertools;
+use parking_lot::RwLock;
+use sketches_ddsketch::DDSketch;
+use std::iter;
+use vortex_error::{VortexExpect, vortex_err, vortex_panic};
+use vortex_expr::ExprRef;
+use vortex_expr::forms::cnf::cnf;
+
+/// The selectivity histogram quantile to use for reordering conjuncts. Where 0 == no rows match.
+const DEFAULT_SELECTIVITY_QUANTILE: f64 = 0.1;
+
+/// A [`FilterExpr`] splits boolean expressions into individual conjunctions, tracks
+/// statistics about selectivity, and uses this information to reorder the evaluation of the
+/// conjunctions in an attempt to minimize the work done.
+pub struct FilterExpr {
+    /// The conjuncts involved in the filter expression.
+    conjuncts: Vec<ExprRef>,
+    /// A histogram of the selectivity of each conjunct.
+    conjunct_selectivity: Vec<RwLock<DDSketch>>,
+    /// The preferred ordering of conjuncts.
+    ordering: RwLock<Vec<usize>>,
+    /// The quantile to use from the selectivity histogram of each conjunct.
+    selectivity_quantile: f64,
+}
+
+impl FilterExpr {
+    pub(crate) fn new(expr: ExprRef) -> Self {
+        let conjuncts = cnf(expr);
+        let num_conjuncts = conjuncts.len();
+        Self {
+            conjuncts,
+            conjunct_selectivity: iter::repeat_with(|| RwLock::new(DDSketch::default()))
+                .take(num_conjuncts)
+                .collect(),
+            // The initial ordering is naive, we could order this by how well we expect each
+            // comparison operator to perform. e.g. == might be more selective than <=? Not obvious.
+            ordering: RwLock::new((0..num_conjuncts).collect()),
+            selectivity_quantile: DEFAULT_SELECTIVITY_QUANTILE,
+        }
+    }
+
+    pub fn conjuncts(&self) -> &[ExprRef] {
+        &self.conjuncts
+    }
+
+    /// Returns the next preferred conjunct to evaluate.
+    pub fn next_conjunct(&self, remaining: &BitVec) -> Option<usize> {
+        let read = self.ordering.read();
+        // Take the first remaining conjunct in the ordered list.
+        read.iter().find(|&idx| remaining[*idx]).copied()
+    }
+
+    /// Report the selectivity of a conjunct, i.e. 0 means no rows matched the predicate.
+    #[allow(clippy::cast_possible_truncation)]
+    pub fn report_selectivity(&self, conjunct_idx: usize, selectivity: f64) {
+        if !(0.0..=1.0).contains(&selectivity) {
+            vortex_panic!(
+                "selectivity {} must be in the range [0.0, 1.0]",
+                selectivity
+            );
+        }
+
+        {
+            let mut histogram = self.conjunct_selectivity[conjunct_idx].write();
+
+            histogram.add(selectivity);
+        }
+
+        let all_selectivity = self
+            .conjunct_selectivity
+            .iter()
+            .map(|histogram| {
+                histogram
+                    .read()
+                    .quantile(self.selectivity_quantile)
+                    .map_err(|e| vortex_err!("{e}")) // Only errors when the quantile is out of range
+                    .vortex_expect("quantile out of range")
+                    // If the sketch is empty, its selectivity is 0.
+                    .unwrap_or_default()
+            })
+            .collect::<Vec<_>>();
+
+        {
+            let ordering = self.ordering.read();
+            if ordering.is_sorted_by_key(|&idx| all_selectivity[idx]) {
+                return;
+            }
+        }
+
+        // Re-sort our conjuncts based on the new statistics.
+        let mut ordering = self.ordering.write();
+        ordering.sort_unstable_by(|&l_idx, &r_idx| {
+            all_selectivity[l_idx]
+                .partial_cmp(&all_selectivity[r_idx])
+                .vortex_expect("Can't compare selectivity values")
+        });
+
+        log::debug!(
+            "Reordered conjuncts based on new selectivity {:?}",
+            ordering
+                .iter()
+                .map(|&idx| format!("({}) => {}", self.conjuncts[idx], all_selectivity[idx]))
+                .join(", ")
+        );
+    }
+}
